+++ CHANGELOG.md
@@ -25,6 +25,10 @@ Add a link to the GitHub diff like
 
 [Unreleased]: https://github.com/mastersign/bench/compare/master...dev
 
+## [0.16.2] - 2017-07-10
+
+[0.16.2]: https://github.com/mastersign/bench/compare/v0.16.1...v0.16.2
+
 ### Added
 - Hints in the transfer dialog of the Bench Dashboard,
   which warn that large SFX files do not work